+++ lib/amnesia.rb
@@ -5,11 +5,11 @@ require 'haml'
 $:<< File.dirname(__FILE__)
 
 require 'amnesia/host'
+require 'amnesia/helpers'
 
 module Amnesia
   class Application < Sinatra::Base
-    SIZE_UNITS = %w[ Bytes KB MB GB TB PB EB ]
-
+    include Amnesia::Helpers
     set :public_folder, File.join(File.dirname(__FILE__), 'amnesia', 'public')
     set :views, File.join(File.dirname(__FILE__), 'amnesia', 'views')
 
@@ -27,39 +27,6 @@ module Amnesia
       user, pass = ENV['AMNESIA_CREDS'].split(':')
       username == user and password == pass
     end if ENV['AMNESIA_CREDS']
-
-    helpers do
-      def graph_url(*data)
-        Gchart.pie(data: data, size: '115x115', bg: 'ffffff00')
-      end
-
-      # https://github.com/rails/rails/blob/fbe335cfe09bf0949edfdf0c4b251f4d081bd5d7/activesupport/lib/active_support/number_helper/number_to_human_size_converter.rb
-      def number_to_human_size(number, precision=1)
-        number, base = Float(number), 1024
-
-        if number.to_i < base
-          "%d %s" % [ number.to_i, SIZE_UNITS.first ]
-        else
-          max = SIZE_UNITS.size - 1
-          exp = (Math.log(number) / Math.log(base)).to_i
-          exp = max if exp > max # avoid overflow for the highest unit
-          result = number / (base**exp)
-          "%.#{precision}f %s" % [ result, SIZE_UNITS[exp] ]
-        end
-      end
-
-      def alive_hosts
-        @alive_hosts ||= @hosts.select(&:alive?)
-      end
-
-      %w[ bytes limit_maxbytes get_hits get_misses cmd_get cmd_set ].each do |stat|
-        class_eval <<-RUBY, __FILE__, __LINE__ + 1
-          def #{stat}_sum
-            alive_hosts.map(&:#{stat}).sum
-          end
-        RUBY
-      end
-    end
 
     get '/' do
       haml :index
